Push common field validation logic to base class
